Switched Array Reduction library to use PtrQ's instead of (shudder) big wonking fixed-sized arrays. In addition to working better (no fixed limit on number of out-of-order messages), the new code is actually *much* simpler!
It simplified the logic so much I felt confident
adding support for element-free PEs.  The idea is to
add a "heartbeat", which parents send their children
when the parent starts a new reduction.  If the children
have already started the reduction, nothing happens;
but if the children haven't started the reduction (because,
for example, they have no elements) this gives them a
chance to tell their parents there will be no contributions
[implemented as a message marked as coming from zero elements].