Major changes to proxies: -Collected registration and callFn definitions into a separate (never-instantiated) "CkIndex_<foo>" object. This improves the entry-index syntax substantially.
-For groups and arrays, created a new proxy type
"CProxyElement_<foo>", which represents a single
element of a group or array.  This eliminates an
overloading ambiguity for groups, provides much
better type safety, is clearer, and eliminates one
"if" in the array implementation.

-All proxies now inherit directly (instead of
virtually) from the corresponding ID type.  This
allows proxies to be sent in messages.  Multiple
inheritence is still supported, by forwarding all
mutator routines up both branches of the inheritance
tree.  Sadly, all proxy methods must be explicitly
overloaded to avoid ambiguity in the multiple
inheritance case.  The type safety, clarity, etc.
definitely make it worthwhile.