Many, many performance changes.
-Factored out datagram-specific code into machine-dgram.c,
in preparation for adding a machine-tcp implementation.

-Made CmiNotifyIdle into a proper spin-then-sleep operation.
This improves the pingpong time slightly on the uniprocessor
version; and is a big win on the SMP version.

-Added a condition variable (or semaphore) for each SMP
processor to sleep on when the processor becomes idle.
This allows the communication thread to wake up a sleeping
PE (by signaling it), rather than just sticking the message
into its incoming queue (as the processor sleeps on).  This
improves the SMP pingpong time by a factor of about 100.

-Removed the writeability check from the CommunicationServer.
Now write packets as fast as the OS will take them, up to
the limit of the send window.  This is needed because a UDP
socket does *NOT* send SIGIO when it becomes writeable, so
there is no quick way to resume sending if we stop.
As a bonus, this results in about half as many calls to
select()/poll() as before, which dramatically improves the
pingpong time.

-Changed acks to a simple "I have all seqnos up to n"
integer instead of sending the entire receive window.
This is smaller, simpler, and much more conducive to
piggybacked ACKs (a la TCP).  With extremely high packet
loss, it may be marginally less efficient (because
higher-seqno packets that actually made it will be resent
along with the missing lower-seqno packet).