Now trap everything going to the terminal, via the odd hack of dup2'ing file descriptors 1 (stdout) and 2 (stderr) as the write end of pipes we open.  This allows us to properly capture printfs, fprintfs, uses of cout and cerr, and even raw "write(2,...)" calls, such as the Fortran runtime generates.
These pipes have to be cleaned out fairly often, though, because a
pipe buffer is only 4K (and fixed in the kernel!).  Even worse, if
the pipe buffer overflows, the last write() call to it hangs until
our SIGALRM cleans out the pipe.  Glibc doesn't reliably recover from
this, so when there are lots of printouts some lines may be dropped.
There is a warning when this may have happened, though.

Tested and works on net-linux and net-linux-smp.  Should work with
gm and tcp versions, although these are untested.  Windows support
is unlikely, although supposedly possible.