Generalized memory layout for hashtable entry.  This is needed because some machines have restictive requirements on alignment, so if, e.g., key is "int" and object is "double", we need a 4-byte gap between the key and object to keep the object field double-aligned.
Without the gap, user code crashes with a confusing SIGBUS.

Also removed ancient, silly hack that "key==-17 -> empty entry".
This saved one byte, but of course silently failed if key actually
*is* -17.  With the alignment problems above fixed, adding an
"empty bit" is pretty cheap and certainly the right thing to do.

In the long run, it may be better to just scrap the typeless
CkHashtable implementation and restrict ourselves to CkHashtableT.
This would let the compiler do *all* the memory layout.