Added MPI_Comm_Split, and the support needed for all sorts of communicators.
A communicator is now implemented as a new shadow array, bound to the
TCharm threads an a "holder" called ampiParent.  This allows the messaging,
broadcasts, and reductions that happen on one communicator to be independent
of other communicators.

This implementation has not been tested well, and there are likely bugs
remaining.  See the FIXME's for details.