Relatively minor portability and migration-correctness fixes:
1.) &(void *)messagePtr isn't actually valid C++, because & only
applies to Lvalues (which the result of a typecast isn't).
You want (void **)&messagePtr, which is valid C++ and will be
accepted by compilers other than the (promiscuous) g++.

2.) Casting addresses to int's doesn't work on machines where
addresses are 8 bytes and ints are 4; you aught to just marshall
"pointer"s instead of "int"s.

3.) CtvStaticDeclare(ArmciVirtualProcessor *, _armci_ptr)
declares a separate _armci_ptr for each file it's included in.
"Static" globals should never, ever go in header files;
if it's used by multiple files, you need CtvExtern in the
header and a CtvDeclare in exactly one source file.

4.) ArmciVirtualProcessor didn't have a destructor, so it leaked
(isomalloc'd!) memory when it migrated.

5.) addressReply isn't initialized to NULL in the constructor,
so pup will crash if you hadn't already done a put.  Similarly,
put didn't reset addressReply to NULL after deleting the message,
so the pup routine wouldn't have worked.