This checkin has the modifications to turn orion's liveViz into a split personality liveViz with both his original methods of running liveViz, and the poll mode.
The code is messy and needs cleanup, but it is there and funcitonal. The
sample server (pollserver) has been developed and tested and checked in
as well.

This update does not contain anything more than the most basic polling mode,
none of the more advanced (so-called liveViz2, liveViz3) layers have been
implemented.