Moved compat_regmm to tcharm library.
This file defines the
        CkRegisterMainModule
that Charm++ uses to start the registration system.
When using -modules, it's not clear this routine is needed,
which is why there's a fallback in compat_regmm.

But blindly defining compat_regmm in ck-core results in
an *empty* Charm++ program linking (and doing nothing!),
which is bad.  Hence compat_regmm is booted out to tcharm.