Added module .dep files.  These files list the command-line arguments needed at link time by a particular module.  For example, a library that depends on some system library libfoo might have a .dep file: -lfoo
Similarly, a module foo that depends on another module bar might say:
-module bar
in the libmodulefoo.dep file.

The .dep file is currently only looked up for modules, and can only
be one line long (for now).  It also only works for modules in the
system charm/lib/ directory, which is just a bug.

This is intended to replace the bizarre language-specific dependencies
currently used for -language fem/femf/ampi/ampif/mblock/mblockf