Added "++runscript" option, which tells charmrun to run the node-program using the script you pass in. (A useful hook for faucets, and probably other queueing systems)
The script will be called on each node with the node-program name and arguments. This even works with ++debug/++debug-no-pause/++in-xterm, in which case the script will get passed a long command string involving xterm.

The runscript must be sure to quote its arguments properly, because it may be called with something like:
	foo "bar baz" baf
The minimal correct runscript is thus:
#!/bin/sh
"$@"