This modification changes the way stream messages are sent between VMI nodes in a computation.
Previously, each stream message prepended a VMI-specific header consisting
of two ints to each message.  This header specified the type of message
being sent (either a normal "short" or "medium" stream message with message
data immediately following the header, or an "RDMA rendezvous" request
message) and the size of the message data the receiver was expected to
handle.

In the new version, no message header is used.  Stream messages containing
"short" or "medium" length data simply contain the message data as before.
RDMA rendezvous requests are implemented via a Converse-level handler which
takes care of publishing a buffer for the sender to place the long message
data into via RDMA.

The changes streamline both sending and receiving routines.  For 0-byte
messages, this results in a savings of 2 us/msg one-way latency (about
15%) and an increase in bandwidth of about 85 MB/second.  There is no
appreciable gain in messages that are transferred via the large message
protocol.