Fixes for MPI_Attr types, used by PETSc.
Though MPI_Keyval's are globally allocated,
the value in an MPI_Attr is local to a single
process' view of a communicator, so the MPI_Attr
values are now stored in the ampiCommStruct.

Also, added support for the builtin MPI keyvals
required by the 1.1 standard, as well as a few
custom AMPI keyvals.

Fortran version of these calls is likely wrong,
and should be tested.