Bug fixes for my most recent changes.
Unfortunately there is a big problem involving either threads or memory
allocation in conjunction with VMI 2.0 Beta 3.  Somehow the stack gets
completely corrupted in a non-obvious way.  Switching to use pthreads in
Converse seems to help this greatly.  It is unknown whether this is related
directly to recent changes in Charm threads, recent changes in Charm memory
allocation (isomalloc?), or whether it's just an unfortunate clash with
changes in Charm + changes in VMI between Beta 2 and Beta 3.

This is going to be painful to fix.