This modification fixes a race condition where two processes can enter CmiSyncSendFn() simultaneously and send RDMA messages to each other.  To do this, they first send rendezvous messages and then VMI_Poll() to wait for the RDMA to complete.  The error in this is that the Converse handler in the rendezvous message never gets a chance to execute.
One fix was to call CsdScheduler(1) in the loop where VMI_Poll() runs.
This caused problems because you could end up cascading multiple calls to
CsdScheduler.

The solution I used was to go back to having CMI_VMI_Stream_Receive_Handler()
check directly to see if the message is a rendezvous setup and handle the
request itself.  This avoids the livelock condition since forward progress
can always happen.

There is a bug still in this code when running Orion's Charm++ marshall test
in megatest.  A fix to that problem is coming.