These modifications fix using spanning tree broadcasts.
THIS COMPLETES MODIFICATIONS TO GET vmi-linux WORKING WITH VMI 2.0 BETA 3.
THIS IS THE KNOWN GOOD VERSION OF THE CODE THAT YOU SHOULD REVERT TO IF
THERE ARE FUTURE PROBLEMS.  This code has passed all Converse and Charm++
tests and has successfully ran NAMD on several molecules up to 64x2
processors.