New delegation interface: a CkDelegateMgr group can now store some data with each delegated proxy in a new "CkDelegateData" class.
CkDelegateData gets attached to the proxy when the proxy is delegated,
gets copied and pup'd with the proxy, and is eventually deleted
when the proxy (and all its copies) are deleted.  Every message the
proxy sends to the delegator includes the CkDelegateData.

This is a suprisingly subtle design problem, because:
  - You want to be able to create, copy, and delete proxies cheaply.
    This means you need reference-counting on the CkDelegateData class.
  - You want CkDelegateMgr to be able to store anything in CkDelegateData.
    This means you need CkDelegateMgr to be involved in CkDelegateData's pup.
  - Proxies are used all over the place in Charm++.

The proximate motivation for this change is Sameer's commlib, but
it's possible CkDelegateData could eventually subsume the CkSectionID
currently used in array sections, as well as being useful elsewhere.