These modifications are a better solution to the problem of spanning tree broadcasts which does not involve forwarding to the spanning children from within CmiGetNonLocal().
The problem before was that a child in the spanning tree would receive an
RDMA broadcast message and attempt to forward the message to its children
in the spanning tree from within the RDMA receive handler.  Unfortunately
VMI never invoked the RDMA send completion handler because an underlying
event on Myrinet (and only Myrinet; everything worked fine on Ethernet and
on Infiniband) never got delievered.  This resulted in the node spinning
forever.

To fix this correctly, I realized that I could set a "disposition" on each
send handle for what to do in the send handler (nothing, free the message,
enqueue the message for delivery to Converse) and then forward the message
to the spanning tree children asynchronously instead of synchronously.
In this scenario, the send completion handler gets invoked after the
receive handler exits, and all is happy.