Added the parallel partitioning code in files parallel_part.C and parallel_part,h. By default the code in parallel_part.C is not compiled. In the default case fem.C doesnt call the parallel partitioning code. So the default compilation of fem shouldnot change.
The parallel partitioning code is activated when FEM is compiled with FEM_PARALLEL_PART.
This activates the code in parallel_part.C and fem.C now calls the parallel partition routine.
In this case FEM needs msa and parmetis. The libmodulefem.dep will be changed to reflect that.
The parallel partition code is not yet in its final shape. The part with multiple ghost layers
needs to be debugged and the rest of the code needs to be optimized and might also need some
refactoring to speed it up.