Task: POSE
There was a performance problem with highly virtualized POSE programs.
It was due to a bug in idle reporting of objects to the GVT.  Early
fixes solved the performance problem but made it impossible to report
an accurate final GVT for certain types of programs.  The fix is to
allow for separate lines of OVT reporting for idle and active objects.
All programs should get a performance benefit from this, particularly
highly virtualized ones (lots of tiny objects).