Added libmoduleseqpose A regular make will create both libmodulepose and libmoduleseqpose. libmoduleseqpose is just pose compiled with -DSEQUENTIAL_POSE=1. Which means you can't simply link via -module seqpose because the module linking system wants various registration functions to exist with the same name as the module.  Attempts to create standalone seqpose modules failed due to the bewildering variety of decl, def, etrans, and pose.h include file namespace agony which result in irreconcilable netsim link failures.
But this cheap hack work as long as you copy it over.
Therefore we also create a libmoduleorigpose copy of the libmodulepose
during the make so we can restore normality later.

The Makefiles for the netsims will have SEQUENTIAL options to take care
of this jiggery pokery (including etrans -s) behind the scenes.
BigNetSim is already done.  HiSim is a bit more complicated.
"Evil hidden from the user is indistinguishable from Good." - Jim Phillips