These changes add a new Charm++ load balancer strategy called RotateLB.
RotateLB is a simple load balancer that rotates the objects on each PE
to the next higher PE, mod the total number of PEs in the computation.
This provides another simple toy load balancer for learning, but is also
a useful way of testing the PUP routines for a Charm++ program.