stripping out the ifdef SPARSE_CONTINUOUS, its been the default for months. the non case has been removed.
Stopped sending the always zero othernewdata in the backward path.
Theoretically this might cause a behavior change somewhere.
But as Glenn's email indicated no use for it, and we're not creating one,
there seems little value in sending it around so its zeros can be
ignored and take up space.

Moved the nogemm calculatePairs to the end of the file.