Checked in the code for the evacuation scheme of fault tolerance. Most of the code is in ckevacuation.C and .h. However there are minor changes all over to make the scheme work.
Also fixed the reduction tree. Removed the ckreductionmgr tree.
Now there is just one tree the node based one.
Allowed ampi to do anytime migration with isomalloc.
Each thread now has a token and it is the token that gets enqueued in the scheduler and
not the thread itself. This is useful for doing any time migration and was a long standing
to do