Much more abundant header comments describing paircalculator.
Refactored creation process to have no redundant code and a more logical
breakdown of creation process.  Trimmed down the makeleft and makeright
functions and made them ready for external calling to assist in the
planned commlib multicast refactoring.

Added support for the new section reduction.
Added initGRed message, the sumDoubleMatrix reduction client,
and the initOneGRed function for the creation of the multicast tree which
will be used by that reduction.

Added proper pup behavior for sparseReducer state.