Added a new gcc4 configuration for the XT3 machine option only.
The XT3 pcqueue hack is apparently no longer required. calloc fails to do
the right job on the XT3 and is currently changed to malloc/memset on the
platform. All other platforms will continue to use calloc.

A decision to use malloc/memset is prefered over the original hack because
it is unclear how the original even helped in the first place, raising
concerns that some bug remains unexercised on the system. malloc/memset
on the other hand behaves semantically the same as calloc though performance
is expected to suffer due to the result of making 2 system calls as opposed
to 1.