partition based on face neighbors rather than node
The graph we were passing to metis was constructed using the rule "any
elements which share a node are neighbors", which produces an
unnecessarily large and low-quality graph. The new default rule is "any
elements which share a face are neighbors". This requires correctly
oriented elements, and does not correctly detect the difference between
quad meshes and tet meshes (we assume tet). To restore the old behavior,
use the flag +Parfum_node_neighbor_graph.