Fix comlib-related build failure under bgxlC
The grand comlib checkin exposed an ICE in the BlueGene version of
xlC. I narrowed the cause in convcomlibstrategy.C to the line in
Strategy::Strategy() that called insertStrategy() to register with the
manager. This commit modifies the manager to 'friend' the entire
Strategy class, rather than just the constructor.

Checked compilation with GCC and xlC on BlueGene, and GCC and ICC on
64-bit lab machines. Things may yet break under the abominable Solaris cc.