TCharm/Emulator: Improve tracing to allow short-cutting library calls
TCharm provides a tracing facility to let developers see when library
code is entered and exited. This commit extends that facility so that
the BigSim emulator logs those entries and exits as events, and makes
the exit event depend on the entrance. Specifically, this lets
BigNetSim detect the call and return of AMPI functions, and enables it
to model them differently than the sequence of messages they generate.