ParFUM: Make MsaHashtable compile against the revised MSA
MSA was revised to return a more strictly-checked object when
accumulating into the array. However, it now passes through overloaded
operator+= on ENTRY objects, to ease some pain of not accumulating the
same type. Define an operator+= appropriate for MsaHashtable, and use
it in place of the older direct access that violated MSA's strict
semantics.