FEM, ParFUM: Account for removed `byte' typedef from IDXL
IDXL provided a "typedef unsigned char byte" up until yesterday. This
was breaking application code, so I removed it. I failed to test that
this wasn't used by things depending on IDXL, but autobuild helpfully
reminded me. Substitute `unsigned char' for `byte' in the handful of
places it appeared. Bemoan the duplication of code from FEM to ParFUM.