Spanning Trees: Overload instead of weird ownership logic
Spanning trees generated for multicasts, reductions, and so forth can
be generated with either a user-supplied strategy or a default one
(selected by a factory knowledgeable of machien characteristics and so
forth). The logic to do this was based on checking if the argument to
buildSpanningTree[Generation] was NULL, and constructing a strategy if
it was. This strategy then had to be deleted afterward only if it was
locally constructed, a somewhat clunky arrangement.

Instead, split the "use the default strategy" logic into an overloaded
function that doesn't take the pointer argument, and passes through to
the (now-simplified) original version. Incidentally, some conditionals
are replaced by compile-time overload resolution.