net-*: Support running on >1024 PEs via poll() instead of select()
select(), unless configured specially, has a limit on number of
sockets (which is 1024). This prevent charmrun from starting a
parallel job greater than 1k cores, for example on ranger.

Use poll instead select whenever it is possible, which is scalable.
If select is the only choice, a CmiAssert is added to detect the
misuse. When fd is greater than FD_SETSIZE (1k default), the behavior
is not defined.