BGP: Update CmiDirect to use newer DCMF API
Release V1R4M1_460_2009-091110P of DCMF adds a few new arguments to
DCMF_Manytomany. This release is now the default on Surveyor, and available on
Intrepid. Update CmiDirect to use the new API, and add a release note about
the need to specify which DCMF driver version to build against.

Ideally, this would have been covered by conditional compilation based on
DCMF_VERSION_*, but those macros are the same under both the old and new
driver versions.