BGP: Make xlc halt on 'Error'-level diagnostics
By default, xlc will report an 'Error' diagnostic for a class of unambiguous
code flaws. However, an 'Error' is not considered fatal. Only 'Severe'
diagnostics actually cause a build to report failure.

An update to DCMF's API caused a function call to have too few arguments,
which xlc was able to flag as an Error. However, because the build still
succeeded, we never noticed. Since nothing else in the Charm code base raises
an Error, tell xlc to treat them as fatal in the future through the '-qhalt=e'
flag.