Don't pass the '-optimize' flag to a hard-coded g++
The charmc compiler wrapper switched from expecting -O to -optimize for the
meaning "pass the best known compiler optimization flags to the underlying
compiler." This was written on the build command line and passed through make
in $OPTS. That works fine as long as charmc gets to rewrite it, but causes
problems when passed directly to a real compiler, like g++.

The file trace-controlPoints-BPG.C needs to be compiled with g++ and not xlC,
and this need was encoded as an explicit invocation of g++ in the BGP-specific
Makefile. Thus, we passed -optimize to it, and got an object file named
'ptimize'. Get this right by using charmc's '-c++' compiler override option
instead.