build util: Script that will generate make dependencies from a ci file
One of the big issues with automating builds for charm++ programs is that decl and
def files are generated during the build. And since the ci file name does not correspond
to the names of the generated headers, pattern rules in make cannot help.

Facilitate the automation of builds by providing a script as part of charm that will generate
make-style dependency information from a ci file. This coupled with existing compiler mechanisms
for C/C++ code should go a long way in making large charm++ projects manageable.