ckcomplex: define operators - and *
Quickly add just the operators required by OpenAtom. However, the real need is to
deprecate the current ckcomplex and provide a substitute to std::complex based on
autconfig tests.

As a side note, there is no charm or openatom code that defines USE_FFTW_DECLS that
ckcomplex checks for. If the other major apps are also in the same boat, then this macro
can be nuked.