SDAG: Reference messages before buffering them, unreference after delivery
CharmLU uses [nokeep] entry methods to make the runtime not make
excess copies of messages that it wasn't going to modify. On
conversion to SDAG, Jonathan and I found that the CmiReference calls
in the client code weren't reached before control returned to the
runtime, because SDAG buffers messages internally when the client code
isn't ready to receive them (i.e. hasn't reached the corresponding
`when' clause).

Since SDAG hangs onto the messages given to it, make it claim a
reference to those messages, and release those references after the
message has been delivered to its client code.