Control Points: Fix potential startup race condition on SMP builds
I was seeing crashes at startup with control points linked in on
net-linux-x86_64. The problem seemed to be that main chares are
started in PE0 in an arbitrary order, so if one of them tries to
access stuff that's initialized in another (such as Ckpv(cp_effects)),
then it will read random bits of memory and fail. When I moved the
call to the initialization function into an initproc call, which will
precede mainchare construction, then that crash disappeared.