LAPI: Only copy argv in PEs that are not rank 0 in their process
This fixes a subtle bug in NAMD in parsing application arguments
together with Charm arguments. NAMD calls converseInit and initCharm
seperately, so the copied argv which will be processed by converseInit
will not be passed to the initCharm, and the argv to initCharm is the
original argv passed by NAMD binary. In this case, NAMD's arguments
are not correctly set.

(cherry-picked from commit 2128c1a42c5ea9db1c4d9d845cb3d19273682cee)