group QHogger to artificially fill the scheduler queue
Xcasts in real applications happen in the presence of other entry methods
in the scheduler queue, and are obviously affected by the q length. Replicate
this context in the test setup in a controlled manner. Now, we can create
scheduler queues of any length with each entry method of specified duration
without any extra messaging on the network. This should help quantify the
degradation caused by having xcast msgs wait in the scheduler queue.

Hopefully, this quantification will point to the need for alternative mechanisms
for relaying xcast msgs:
    - assign xcast msgs higher priorities?
    - write special converse handlers that will relay xcasts to children
      before hitting the scheduler q for local delivery?