Spanning trees: Dont count on-node subtree against maxBranches
Strategies that try to minimize the total bytes on the network use one sub-tree
for the on-node PEs. This used to count against the maximum allowed branching
factor specified for the spanning tree. However, in the case of binary trees
(maxBranches=2), this caused excessively deep trees as the inter-node tree was
effectively a chain! Fix this by not counting the local branch against the
allowed maximum branching factor (maxBranches).

As before, if there are not enough off-node children, the strategies will use
as many branches for local destinations as possible.