dimInfo: Explicitly default to nchares = 0
When an array is constructed with no initial elements, DefaultArrayMap
sees that nInts == 0 and passes through to RRMap. However, dimInfo
still does calculations based on the initial arguments. These
calculations trigger valgrind to note calculations based on the
uninitialized value of _numChares, which could legitimately be
garbage.

Initialize _numChares to 0 in the case where it won't be used, to
quiet the errors.