Fix allocation/initialization bug in messages with varsize arrays
Store the offsets computed during message allocation and assign the
pointers to later in the buffer during the constructor of the
message's generated base class. This saves us from cases where the
compiler overwrites the buffer's contents between allocation and
construction.

This patch does not address the cases in which
 - The compiler generates code in the user's constructor that
   overwrites these pointers
 - Alignment is screwed up by the compiler padding the allocation at
   the beginning rather than the end.
We don't currently know of instances in which either of these errors
occurs.