Make.lb: List LBs that need external libs for make depends, etc
Don't compile MetisLB, ScotchLB, and TeamLB as part of EveryLB, but
still list their various pieces in the generated Makefile.

When running 'make depends', their source files will need to be edited
to leave out the external headers. This should be a much rarer
operation than people wanting to build with those, which is being made
much easier.