ckcallback: Use a CkArrayIndex base class object instead of CkArrayIndexStruct
CkCallback has a union of data based on the different callback types. One of
these if for callbacks that go to array elements. This required storing an
array index in the union with the resulting stipulation that the stored index
object should not have any default constructors.  A CkArrayIndexStruct class
has existed in the codebase expressly for such situations. Because this class
has no relation to the hierarchy of 'normal' array index classes, dangerous
pointer casts have been the norm.

Because of such risky casts, this snippet of code in the callback has been
plagued with heisenbugs over many years, as compilers get stricter and exhibit
greater personality variations across brands and architectures. OpenAtom has
exposed the bugs a couple of times in the last couple of years.

To fix this issue for good (hopefully), the objective is to not cast across
unrelated class hierarchies, but rather make the base class amenable to
participation in a union.  This should help get rid of classes that existed
purely for such silly hacks to enable charm functionality.