xlat: Generate redn_wrapper fn definitions for mainchare redn targets too
Mystifyingly, the typed reductions example code would not build because the
generated code contained declarations, but no definitions, of the redn_wrapper
functions. It turns out this behavior could be toggled by changing the
reduction root from a mainchare to a regular chare. Some investigation with
Phil's help seemed to indicate that the genCall to generate the redn_wrapper
definition could not be reached for mainchares because of an incomplete if
condition. Fix.