charmxi: Factor out duplicated code calling `when' bodies, fixing bugs
The code that unpacked state variables to call the body of SDAG `when'
blocks was essentially duplicated across cases with 1 and more than 1
clause. One branch of the duplication had been modified in cfb58ab1ad
to fix a bug, but not the other one; the consolidation fixes
that. Also address a silly memory leak in the process.