charmxi: Parse entry method template declarations in .ci and emit accordingly
All of the generated code in .decl.h/.def.h files are templated
according to whether the corresponding entry method is templated.
Push #ifdef CK_TEMPLATES_ONLY down to each element, so it can switch
on whether it's templated or not.

No syntactic support for instantiating entry method templates yet,
though it can be done manually.