Bigsim bare emulator examples: provide a dummy version of traceWriteSTS
Commit 1e05b13735befbeb78ba3ac9743f60e40047f279 made the Bigsim
emulator generate and STS file automatically, to remove the need for a
manual execution with Projections linked. That assumed that the
program running under emulation was going to link the Charm++
libraries, which doesn't hold for this small set of examples. Since an
STS file is meaningless for them (I think), provide a dummy version of
the missing function.