Chare Arrays: Add interface to signal online dynamic element insertion
The AMR benchmark code was having trouble with hangs around
AtSync-triggered load balancing, apparently related to the fact that
it was inserting new chare array elements after startup. In
discussions with Gengbin, we figured that this arose because the
additional elements were interfering with the count of local elements
used to determine when a processor's objects had all reached AtSync
and thus it could contribute its LB stats.

The backend infrastructure had interfaces to 'pause' the LB barrier
while new elements were being inserted, but this had never been
surfaced in a way usable by client code. Add a `beginInserting()`
method as an analogue to the existing `doneInserting()` method that
was already required at startup to surface that code.

Tested to resolve at least one kind of hang in the AMR code on BG/Q.