Revert tracking chare constructor definitions in the grammar
We should do it in the code-generation instead, when we can do a sane check and
emit a sensible error message.

Additionally, the reverted code was crashing on the type of error it was
supposed to detect.

Revert "charmxi: track constructors in chare arrays and templated types"
(Partially) Revert "charmxi now drives home the error message"

This reverts commit b220b31ad4ea31647859bb724472caf7efbe5cdf.
This reverts commit 7005ddd7b081bbd066093bb4f1adb2d371bdab3c.