SDAG: ensure that necessary data structure is initialized
The __cDep structure that SDAG depends on to track the dependencies and state
was not being initialized in some edge cases. Fix them by having every declared
SDAG entry method (either top-level or when-defined) check and initialize this
if it's missing.

In the case of a chare class constructor calling some SDAG method on itself
from the constructor, the surrounding generated method's call to
impl_obj->_sdag_init() won't have happened yet. In the case of inheritance from
a class with SDAG code, the generated code that would have called _sdag_init()
was never running. Incidentally, this may enable SDAG to function for classes
with SDAG content at multiple levels of their inheritance hierarchy.

The change to auto_ptr is necessary because the bare pointer was being
value-initialized in some cases, and could thus be non-NULL junk that would
pass the test.