Use the new dynamic CPU set allocation API to support very large SMP machines
In particular, allow a multicore binary built on Stampede to run on
Blacklight.  (It's not needed if you build on Blacklight, because the
default CPU set size in the headers has been increased on that
machine.) This should be safe, as it doesn't use the new interface
unless the default size is too small.