partition: adding support for user defined partitioning.
If a user wants to use his own partitioning scheme, he should write a C
function with the following syntax:

extern "C" void createCustomPartitions(int numparts, int *partitionSize, int *nodeMap);

numparts (input) - number of partitions to be created
partitonSize (input) - size of each partition
nodeMap (output, prealocated) - is a preallocated array of length - CmiNumNodesGlobal.
Entry i in this array specifies the new noderank original node i should assume in the
mapped space. The entries in this array are blockwise divided to be as the new partitions, i.e
entries 0 to partitionSize[0]-1 belong to partition 1, partitonSize[0] to
partitionSize[0]+partitionSize[1]-1 to partition 2 and so on.

Also, to charmc, one should pass -custom-part, in absence of which a default function
is linked (which will abort the program if called).

As a result of this addition, make.depends, and Makefile are also modified.