SDAG: fix long-standing code generation bug for "when foo(...), foo(...)"
If whens are connected by a comma (AND) and the name is the same, then the code
being output was causing a name collision. This was simple to fix, but actually
fetching multiple messages from the buffered messages requires selecting
different ones (with possibly different reference number constraints). Solve
this by adding a set<CMsgBuffer*> of pending found matches and ignoring those
when searching to another message.