charmrun: use bash instead of sh for the remote shell script
A strict POSIX implementation of sh doesn't need to implement command line
arguments for the '.' builtin (aka 'source' in bash). This makes it harder to
write a '.charmrunrc'. The specific example on which this causes trouble is an
Intel compiler environment setup script that expects the target architecture as
a command line argument.

Change-Id: I61be2978f838ac960a98510f632c6fbd6e4e04b5