Bug #480: Make migrateMe() more forgiving by executing it asynchronously
The API routine `void migrateMe(int toPe)' has historically been very
unforgiving, in that any member access after it's called could result in
undefined behavior (reading/writing members of an object that's been
deleted). This is particularly pernicious in combination with SDAG code, in
which the user may not have written any code after the call, but the generated
code has to touch member fields to implement subsequent control flow.

There's nothing really fundamental about running this method perfectly
synchronously - the important bit is that the object migrate as ordered with no
further direction or runtime system mediation. So, add a bare entry method
ckEmigrate() that does only the inline emigration call, and nothing else, and
turn migrateMe() into an asynchronous invocation of that.

Co-authored-by: Evan Ramos <evan@hpccharm.com>

Change-Id: I7a200b7232bf39c5619fe2bc7f62c5097dc7ff03