charmxi: Always emit a definition of the ChareClass_SDAG_CODE macro
Rob Van Der Wijngaart of Intel noted that the definition of this macro
being conditional on the presence of SDAG code in a given chare class's
declaration was a usability impediment. Specifically, removing some
unused SDAG code from the .ci file could cause compilation failures in
the C++ source files.

Change-Id: I5d60b2dee2b2d8f9ba67c48ec1a8283c7d75ec48