Standardize make commands to use "$(MAKE) -C"
Previously these commands were using a variety of make and $(MAKE).
$(MAKE) is preferred for recursive calls because it will use the version
of make used to invoke the top level call. This also passes the
MAKEFLAGS variable for sub-makes. See section 5.7.1 of the GNU make
manual for more info.

Another form of variation was the use of "cd DIR ; make", which is
will not correctly catch errors, or "cd DIR && make", which will correctly
catch errors. I have used the -C flag instead per the instructions in
section 5.7 of the GNU make manual on recursive use of make. This will
generate the equivalent of "cd DIR && make" which will fail if the cd
fails instead of not cd'ing and silently continuing.

Change-Id: I139231172c7422ac556a5a81f734f541173746d0