Array reductions: When deleting the whole array, stop counting contributors
When a chare array is being destroyed, the infrastructure to count contributors
on each PE, and restructure the reduction tree when one becomes barren, is
just extra work.

This extra work causes a problem because it sends messages to group elements
that are being deleted. These messages represent leaked memory each time an
array is destroyed. They also prevent QD from ever triggering thereafter.

Discovered and characterized by Tom Quinn in discussion on
http://charm.cs.illinois.edu/gerrit/433/

Change-Id: I04a7c994a08fdc0485c344157ac872287e5e1f21