Fixing an interface issue of collision library
Charm collision library provides an interface
for a "client" group to be defined but it has
to extend a client group it provides. Therefore,
when the library is built, its .dec.h header
needs to be copied to /include. This change addresses this.
A mpi-interop example was hacking around by
copying that declaration file, which is fixed here as well.

Change-Id: I72c80ca8048f98fafc1e3941c60c620eb7214c04