Cleanup #600: array4d test not run in nightly build.
Test contained several bugs.  The worst of which was improper use of
CkMulticast for marshalled messages, which caused a hang under some
conditions.  As there is no other example of 4D array use, I fixed the
bugs, removed the pointless and incorrect use of CkMulticast, revised
the CkArrayMap to map using all 4 dimensions with z innermost, and
added a README.  Given that it isn't a very an interesting test, but
is a decent example, I moved it to examples/charm++/hello and renamed
it to match the multi-dimensional array directory naming scheme there.

Change-Id: I167c462230b6e6abb60cf0e807301778d20527d3