ckio: Disable currently irrelevant assertion
CkIO currently doesn't do anything with the options that control which PEs it's
supposed to use. That makes assertions about their validity irrelevant.

The defaults for CkIO's PE placement options are weak, but they're currently
not hurting anything. So, don't try to constrain them in ways that will break
simple functional testing.

Change-Id: I122063257fc0d34befde40e277364147c2f6debb