Feature #420: Set ppn to be 1 by default
A natural way to compute the number of processes per physical node is to
divide the number of PEs by the ppn, which does not work by default
since ppn is initially set to zero in non-local and non-scyld cases.
Here we change that to be one, which is functionally equivalent but
numerically more convenient.

Change-Id: I50bf4a0560c487a9577665fdd5993cbb3a8c8511