Bug #745: verbs-smp performance problem
Broadcasts were based on point to point instead of hypercube or
spanning tree. Remove the lines in config where spanning tree and
hypercube are set to zero. Spanning tree is enabled by
machine-common-core. Remove the buffering of the broadcast messages,
and hand them over to lrts directly. Otherwise, it causes a hang.

Change-Id: I960022067d4cd9a400fdad51da7a6997ed08624c