Bug 619 liveViz does not work with 1 processor
The bug is because the scheduler does not check if there are any incoming
messages if there is only 1 processor. The fix is to add a flag to notify
the scheduler when there is a ccs message and when running on 1 processor.
The scheduler will check the incoming message queue in that case.

Change-Id: Ifb024e79cfec712b6f3bc86681c3e4196b07db18