Add CmiEnforce macro to check conditions in production builds
Implement CmiEnforce macro, patterned on CmiAssert, but always on.

For verbs/ibverbs substitute CmiEnforce for CmiAssert(0) everywhere and
for CmiAssert tests for error conditions during startup.

Also refactor __cmi_assert to pass error message to CmiAbort so that
charmrun might forward it to user, and make ibverbs charmrun do so.

Change-Id: I3cd91b5d6ff8ea6133561b334594507a48cbdc8e