NO MERGE YET - Extend TopoManager to provide ordered list.
FEEDBACK REQUESTED

TopoManager_createPartitions is being extended to get an ordered list of nodes
for easy partitioning of available nodes. TopoManager_createPePartitions is
being added as a convenience function if PE list is desired, though the size
of partitions is still determined by the number of Charm++ processes/nodes.
Blockeed traversal scheme is being copied over from OpenAtom; OA will be
modified to use this API.

Change-Id: If0ae535ce6120bdeb13b2066d2d35b4798a60102