Bug #1062 / Change 1189 followup for non-LRTS: Fix reference to undefined variable
Fixes linking errors on NET and PAMI machine layers, which we aren't quite ready
to discard. The example in `examples/charm++/user-driven-interop/`
compiles cleanly on netlrts, and fails to link on net, as somewhat intended.
It would be a bit nicer if we could give a nice compile-time error, but this
at least un-breaks the build without creating any more pathological conditions.

Change-Id: I5d500deee04691e174f791ec74fe626733b2bda6