AMPI #1193: do not assume all user-defined Ops are non-commutative
Change MPI_Op from a function pointer to an int, used as an index into
a table of Ops. The table also keeps track of commutativity.

Reductions on commutative user-defined ops pack an AmpiOpHeader into
each contribution. Reductions on non-commutative ones do not: instead we
keep the MPI_Op in the RednReq and use that to correctly process the
contributions at the root.

Change-Id: Ib954d0968c518fbce9a1c93f65c6c30978491392