nodeGPU: Scale GPU Manager host mempool by node size
Currently we allocate a pool on the host via cudaHostAlloc. Each instance of
GPU manager allocates a pool equal to the size of the GPU's memory divided by
the number of GPU manager instances on that node. Thus, the total pinned memory
is always equal to the size of the GPU's memory. We can scale this total by a
defined constant (default is 1). This memory is chunked up into separate buffer
pools of doubling size (256, 512, 1024, etc bytes). The number of each buffer is
calculated using an array of relative ratios. We pin this memory in blocks for
each buffer pool size after cputopology is set up so we can use it to poll
physical and logical node sizes.

Change-Id: I08615a9f60dc1bf928e48993ebfee4b91be27bd8