Bug #1323 megatest multisectiontest: Avoid race between array construction and use
The test tries delegating a section proxy over a bunch of arrays. Make that
test depend on construction of the last of those arrays before it starts. This
isn't 100% robust, since the arrays themselves could be constructed out of
order (particularly, the last one in program order before at least one of the
others), but it should be much more reliable than before.

Change-Id: Ie33139d6e8669f6bf1d6746780decd403d38aaa7