CentralLB: added 'concurrent' mode
In concurrent mode, every PE runs the centralized
load balancing algorithm and calculates a solution.
The best solution has to be selected, and subsequently
applied at the PE that generated the solution.

To use concurrent mode:
- Subclass of CentralLB has to set concurrent=true.
- CentralLB will send the LB stats to every PE, thus calling
the LB strategy on every PE.
- Subclass of CentralLB has to implement the mechanism
by which the best solution is chosen. When the best
solution is known, the subclass has to call CentralLB::ApplyDecision()
entry method on the PE that generated the best solution.
Migrate message will be created only on this PE.
- ApplyDecision then causes CentralLB to apply the LB decision
as usual.

If concurrent=false CentralLB works exactly as before.

Change-Id: I479f7f1a5b1d21060f3c646a1564ca4d0ea7383e