Feature #969: AMPI support for inter-communicator (i)gather collectives.
This implements a local (i)gather operation to an intermediate root
in the remote group ranks, which is followed by a single send of the
accumulated data to the root in local group.

This patch also removes ampiCommStruct::isInter. The check for an inter-
communicator is now done by checking if ampiCommStruct::localComm is
MPI_COMM_NULL or not.

Change-Id: I25fc1695f04437fe411ca096dbf95b7d8db9b1b4