Bug #337 TopoManager: query cputopology when topo info not present
Before, if no machine-specific topology information was present,
TopoManager would consider each PE to be a separate physical
node. Now, it will query cputopology (converse module) if
no topology info is present.

What cputopology does in this situation is to group PEs based on
the IP addresses of their host, which are sent to PE0 in a
converse-level reduction.

Note that if cputopology can't obtain socket addresses in a
particular system, it will at least group PEs by logical node,
which is still useful.

Change-Id: If0be00be31b7bdbaacf9d24bd045610a2887ecba