Bug #1278: Make CkReductionMgr node aware for SMP array and group reductions
Instead of duplicating more logic between CkReductionMgr and
CkNodeReductionMgr, we changed the spanning tree construction for
CkReductionMgr so that CkReductionMsgs get aggregated at the node level,
although for now this means all of the reduction work happens only on
rank 0 of each node. This removes the need for CkReductionMgr to
interleave with CkNodeReductionMgr and completely removes the need for
CkArrayReductionMgr. The duplication removal is related to Cleanup #12.

Change-Id: I4b98fa8348c20b63bf5149782262506c09316c30