Bug #1424: Simplify message queue structure when it's randomized
Replace the complicated nest of a priority queue of buckets of
messages, that needed to be iterated over to select one at random,
with a simple flat vector from which random selection is trivial.

Remove some unused member functions that would have been troublesome
to implement for the new version of the randomized case.

Change-Id: Ia7f6d7733691dff3dcc531d679fcee3500de5dbb