Automatically incorporate TRAM when using [aggregate] entry methods
In Foo.decl.h, #include <NDMeshStreamer.h> when any entry
method on a chare in module Foo is marked as [aggregate].

In Foo.def.h, call the module registration for NDMeshStreamer from
module Foo's registration.

Offer the NDMeshStreamer library and its completion library dependency
for linking. Written as -lmoduleNDMeshStreamer instead of -module
NDMeshStreamer so that its registration isn't automatically stuck in
the main module generated code, and thus none of the object files will
get pulled in if it's not in use.

Change-Id: Ic104f5659839e5d240cdf844d7401ca51ec8449c