Bug #1493: Add check to prevent bad deletion of bound array elements
This changes the deletion path of array elements so that CkLocRecs only
get deleted when there are no longer any array elements bound to them.
The key change in the deletion flow is that instead of ~CkLocRec()
ultimately resulting in ~ArrayElement(), ~ArrayElement() calls
CkLocRec::destroy(), which only conditionally deletes the location
record from within it's location manager.

Change-Id: I48c6c3a947a97be233b0476d87aba898ecac5778