Cleanup #1432: Document LLVM OpenMP runtime integration
Documentation of LLVM OpenMP runtime integration with some changes to
build script, charmc and enviroment routines to decide number of
threads.

-fopenmp or -openmp is included in the predefined compiler flags.

Users can limit the number of OpenMP threads on each chare by
'OMP_NUM_THREADS and omp_set_num_threads'. The OpenMP runtime decides
the number of threads for each OpenMP region not to exceed this number.

Specified 'CmiGetCurKnownOmpThreads' for the use case of shared array
between threads in OpenMP.

Added an example to show how to use the integration.

Change-Id: If4d66e4a96656989663438123b7c21c6cfda3bef