AMPI: implement (I)Ssend's using a synchronous rendezvous protocol
Instead of sending the message payload eagerly and then getting an ack
when the recv'er actually matches it, now we use a rendezvous protocol,
only sending the payload after it has been matched.

The initial sync message (to ampi::genericSync) has no payload and its
refnum is the index of the sender's request. When the sync message is
matched by the recv'er, we delete that placeholder msg and send a
message back to the sender (to ampi::ssendAck) to get the payload back
in a normal AmpiMsg (via ampi::generic or ampi::genericRdma).

Change-Id: I6a8db194f63321a2247f707042d624d39fa3bef8