AMPI: use Ssend protocol for PE-local sends to avoid intermediate copy
PE-local sends are done inline if the message is in-order and expected.
If not in-order and expected, then block the sender on the matching recv.

Build AMPI with -DAMPI_LOCAL_IMPL=0 to disable all local/inline sends.

Build AMPI with -DAMPI_LOCAL_THRESHOLD_DEFAULT=N and/or set the
environment variable AMPI_LOCAL_THRESHOLD=N, where N is the number
of bytes below which AMPI will not block the sender.

Change-Id: I244e400fa8e8750c2a999109572ca12124f4d672