AMPI: use Ssend protocol for Node-local sends to avoid intermediate copy
In the payload of the Ssend message we serialize an SsendInfo object
that contains the sender's information, which can be used by the recv'er
if in the same logical node (process) to directly copy the real message
payload directly from the src buffer.

Also change the existing AMPI_LOCAL_THRESHOLD to AMPI_PE_LOCAL_THRESHOLD
to distinguish it from the new AMPI_NODE_LOCAL_THRESHOLD variable.

This means that all pt2pt messages with size greater than or equal to
the AMPI_NODE_LOCAL_THRESHOLD are now 'zero-copy' if in the same node.

Change-Id: Iebe9972b630fe17f4cc5c2a7ae0aa9f4b54b295b