charmxi #1614: Fix SDAG closure structure pup method generation for rdma parameters
This bug was caught in the nightly build for target mpi-linux-x86_64-smp on respect(machine).
The crash happens on unpacking and using the rdma buffer after the closure structure
on the new PE after migration. The fix was to adjust the rdma pointers before and after
the migration in the pup method of the closure structure generated by charmxi.

Change-Id: I83a45a41f5cc2140632cbcd0d23631046dc90b39