TopoManager: Added API to query physical node info
TopoManager can build and store list of physical nodes and their
coordinates according to machine topology, and generate the adjacency
lists of nodes in network graph.

Information that can be queried for physical nodes includes the set of PEs
in the node, coordinates of the node, and neighbors of node in topology.

Also added method to TopoManager to query allocation shape, which works
for any general topology, including Blue Waters allocations with holes.

Change-Id: Idc7d27fa763be3454c671d87ba4ae8a464ac8862