AMPI: maintain a pool of commonly used Request objects
AmpiRequestPool is owned by the ampiParent, and is used for IReqs,
SendReqs, and SsendReqs. It optimizes for the common cases of:
- Iteratively creating and destroying a small number of requests (<=64)
- Iteration over a mix of send and recv requests, as in MPI_Waitall

Users can set the size of the request pool (the number of requests in
it) at build-time with '-DAMPI_REQ_POOL_SIZE=<size>".

Also change MPI_I_REQ and such to typed enums and rename to AMPI_I_REQ.

Change-Id: I460a098405d5034679c93868cc2314c05405811d