TCharm: try to avoid over-allocating Isomalloc stacks with power-of-two sizes
Isomalloc uses the mempool to allocate memory in power-of-two sizes.
TCharm stack sizes are almost alwasy given as even power-of-two sizes,
but Isomalloc and mempool insert their own metadata in the allocation,
pushing the actual allocation size to the next power-of-two size up.
So for the default stack of size 1 MB, we were actually allocating 2 MB.

Change-Id: I06c408fad0a07b3e03a511b8d4be6a675e1b70dc