HierarchicalLB: implement scalable hierarchical LB strategy.
The implementation is directly based on the description in this paper and the
implementation in TASCEL:

HPDC'12
"Work Stealing and Persistence-based Load Balancers for Iterative Overdecomposed
Applications"
Jonathan Lifflander, Sriram Krishnamoorthy, Laxmikant V. Kale

In some preliminary tests with SimplePIC, it performs much better on 2k cores
than HybridLB (2 seconds vs. 0.14s). It is slightly slower than DistributedLB,
but provides a much better balance. It is much faster than HybridLB.

The algorithm using a histogramming/sampling method to bin work units. It builds
a specialized tree (an n-ary two-rung tree) to perform balancing, moving up and
down the tree to distribute load based on node averages. It is greedy in that it
does not attempt to move objects that belong to very underloaded proccessors.

Although it is performing well (better than DistributedLB and HybridLB for the
tests) counting both balance quality and final iteration time, there are several
major opportunities for optimization: Currently, the distribution algorithm
iterates through each element when selecting a min node when assigning
load. This can be optimized by moving elements in bulk during the local
distribution phases.

Meta-information is tracked using a 64-bit identifier scheme. The first 32 bits
represent the PE, the latter 32-bits storing a dense index into that local PE's
object map.

Change-Id: I5f17308127c6b33d3cd92f1cf4c1b93bd6390d83