AMPI: use zero-copy direct API to avoid intermedate copies
For messages sent within the same PE or process, we pack an
AmpiNcpyShmSource object into the AmpiMsg data. This allows direct
handling of DDTs in a single message, and avoids the cost of memory
pinning when it is unnecessary. We still have to handle the case where
the recv'er turns out to not be in the same process, so the recv'er
then requests an rput().

For messages sent across processes, we pack a CkNcpySource object into
the AmpiMsg data and use that to invoke an rget() from the recv'er.

Change-Id: I9e8126bc8bf321b4a5ab64e08b3283e448fec4c0