Bug #1770: avoid build-line options on host compiles
Set CMK_NATIVE defaults before reading conv-mach-opt.sh to prevent
target-specific build-line options such as -xMIC-AVX512 from
being used when building host tools such as charmxi, which then
fail when run later on the cluster head node, for example.

This is a return to the order used in Charm++ 6.8.2 and earlier.

Change-Id: I501a79552ac7392996b8431e4df5d22ff04b528c