AMPI: store AmpiRequest pointers directly in the posted request queue
Previously, we stored MPI_Requests (ints) in the posted request queue,
and used those to look up the AmpiRequest objects through the
AmpiRequestList. This change removes that indirection by storing the
object pointers directly in the posted request queue.

The only complication arises during migration of AMPI ranks with pending
or posted requests, since AmpiRequest objects are owned by the
AmpiRequestList, not by postedReqs. So we add 'reqIdx' as a member of
the AmpiRequest base class, pack the reqIdx when migrating and use that
to find the AmpiRequest* when unpacking in ampiParent::registerAmpi().

Also rename the message matching queues from 'msgs' and 'posted_ireqs'
to 'unexpectedMsgs' and 'postedReqs'.

Change-Id: I077f0bb37984b1b328a544852a99be4d4d85c587