Fixed CmiNodeBarrier() for PAPI Initialization:
Initial implementation of CmiNodeBarrierCount() assumed a NodeAllBarrier
will complete before a NodeBarrier is called (and vice versa), however
the workers in initPAPI() need to call NodeBarrier before they can complete
the NodeAllBarrier previously called by comm threads. Original
CmiNodeBarrierCount could not handle this situation with one set of
synchronization variables, so I added a second copy of everything, allowing
CmiNodeBarrierCount to handle both barrier types at the same time.

The pairs of elements are stored as the first and last elements of arrays
that are large enough that the two elements will exist onseparate cache
lines as to avoid false sharing.

Change-Id: I3a4c154adb391f8e2361a53fa8411c749c720fb5