Bug #1832: SMP hangs in megatest/multisectiontest
This fixes the netlrts-linux-x86_64-smp hangs seen in nightly build
and jenkins. The bug was caused because of updating the ‘mProcessed’ QD
counter twice for chare and nodegroup messages that were buffered for not
satisifying group dependence requirement. This led to failed QD, leading to hangs
in tests where QD was supposed to be detected. The bug manifested in the form
of a hang in the multisection test because of incorrect QD counter
processing in the previously executed groupdependence test.

Change-Id: I71e49128f0dc488afbebca3b9ac13d66295d5333