Bug #1852: Fix verbs-linux-x86_64 failure after zerocopy direct api merge
The failure occured in tests/charm++/pingpong as a part of the ack handling
process of the old Rdma API, called CmiDirect. Since this code was touched
by the Direct API, the failure was caused. The fix was to revert the previous
behavior for CmiDirect_Put by not performing any ack handling on the sender
after a CmiDirect_Put call.

Change-Id: I1ce338b2d03155f31d604f188a959907a44bb7a6