netlrts TCP performance optimization using sendmsg
This fix substantially improves the performance of the TCP layer.

The TCP layer disables Nagle's algorithm (with TCP_NODELAY), which
reduces latency and seems like the best option given benchmark
results. However, a consequence of disabling it is that packets
are sent immediately on calling send(socket,...).

Previously, the TCP layer would call send on a socket two times
for each message, one to send the message length (one int),
and another to send the actual message. The combination
with TCP_NODELAY is horrible.

The issue is solved by using the 'sendmsg' system call, to which
you can pass an array of buffers and the OS puts all in one TCP msg.

Note: another big performance boost can be obtained by setting
FRAGMENTATION=0 in machine-tcp.c for large msg sizes.
However, this seems to lead to issues when sending huge messages
(like with AMPI migration of virtual ranks). Not sure what
FRAGMENTATION does or why it's needed, but it can be something to
look at in the future.

Change-Id: I04f7e58f3a06daa8ccfd10d598c72341b882fa29