Bug #1947: Fix mpi-win-x86_64-smp hang for nodegroup zerocopy api in pingpong
The hang was because MPI was initialized in MPI_THREAD_SINGLE mode,
but instead of the comm. thread, the worker thread was posting the
recv buffer. The fix was to have the comm thread post the buffer instead of
the worker thread.

This commit also includes another fix related to the zerocopy api used
for nodegroups. For nodegroups, the pingpong example crashed for
mpi-win-x86_64-smp when run on 1 PE because of a stack overflow. This
was due to inlining of the send method causing recursive calls which
caused stack overflow when run with a large number of iterations (~900).
The fix was to not use inlining in this code path.

Change-Id: I1c86dbac2ce81832d5f728aadf769fe9e49cf9f7