AMPI: Fixed bugs with predefined datatypes
Changed the serialize functions to pass in the length of the
message received, as well as the length of the buffer, so in the case
the case that the buffer is longer than the message it only
reads up to the length of the message.

Serialize now writes even if only a full datatype was not sent.

Reordered the values for MPI datatypes (MPI_BYTE, CkDDT_INT, MPI_DOUBLE_INT,
etc) so that simples datatypes containing one value such as MPI_BYTE are below
structs such as MPI_DOUBLE_INT, and added CkDDT_MAX_BASIC_TYPE to divide them

MPI Structs such as MPI_DOUBLE_INT are now treated as CkDDT_Struct
instead of CkDDT_DataType so that it can store the locations of any
datatypes it contains for use with serialize

MPI_Get_elements now returns the number of basic elements instead of
the number of elements, and even if only a partial datatype was sent.
Added a function getNumBasicElements() to help with that.

Reworked Indexed and Indexed_Block to inherit constructors from
HIndexed and HIndexed_Block respectively since their code is very similar.
To do this added an extra parameter to the Indexed and Indexed_Block and
constructors that takes an array of displacements in bytes to feed into
the respective  HIndexed and HIndexed_Block constructor.

Indexed_block now takes an array of ints for the displacements as the MPI
stand and MPICH both specify that it should receive an array of ints

Fixes tested in MPICH test cases get-elements-pairtype

Also fixes segfault in mpich/coll/bcast_comm_world by checking in
MPI_Type_set_name if the type is NULL before attempting to set the name

Replaces CkDDT_ constants with MPI_ constants

Change-Id: I200c81fdba8045a385892a001fda509e2e2373cd