AMPI: use concat rather than set reducer for gather(v)-like reductions
Use concat for gather, gatherv, and noncommutative reductions. The set
reducer adds unnecessary memory overhead compared to concat, since each
contribution has an int size along with it. For gather's and
noncommutative reductions this is straightforward because the
contribution sizes are all the same, while with gatherv's we add a
separate tuple reduction to keep track of contribution sizes.

Change-Id: Ied154413696c140e4a25325bb7787cab1febaed5