Fix TCharm applications' renaming of entryInfos
Always store the demangled name for templated entry methods, to
simplify the state of name (it no longer depends on tracing being
enabled), add a bool to entryInfo to track ownership, and add a setName
method for users that want to rename entryInfos. This removes the
allocations in the non-templated, non-renamed common case.

Co-authored-by: Evan Ramos <evan@hpccharm.com>

Change-Id: I066b8795343f0c484da04d6e0726c6fdaf79e14c