Bug #2037: Execute callbacks to inline EMs in an inline manner
Previously, callbacks targeting inline entry methods had to be
declared with a 'true' value for the doInline variable to specify
inline execution. With this fix, callbacks execute inline when
the target is an inline entry method. The CkCallback constructors
still have the 'doInline' variable, now called 'forceInline', to
preserve existing behavior.

Co-authored-by: Evan Ramos <evan@hpccharm.com>

Change-Id: I2b21c7848b2428a367e79cdbaa9424f929c27ee1