Rename the majority of remaining C files in the RTS to C++
It is unlikely that this patch will pass verification, but all problems
should be fixed in subsequent patches merged at the same time, and I
prefer to keep the patches separate for better historical organization.

If this commit causes problems with git bisect, use a bisection script
and ensure it emits error code 125 when a build is unable to complete.
See https://git-scm.com/docs/git-bisect#_bisect_run for more
information.

Change-Id: Iaaa03b3436ea4b2a69b4f4e5035e146be2199530