ZC Direct API: Reset counter immediately in pingpong examples
Previously, the counter was reset inside endGetGetPingpong and this
caused an occasional hang in the execution because the incremented counter
value never reached the desired value. This issue is fixed by resetting
the counter immediately inside the beginGetGetPingpong method. This
patch fixes the multicore-linux-x86_64 hang seen in last night's
autobuild(29th April, 2019).

Change-Id: Iedc460aeb11438f06945f3b2697879bf17cfcf86