ZC Bcast API: Fix hang in smp mode because of unsafe thread access
Previously, the variable 'counter', which keeps a track of the number of child
Rgets completed was being incremented by both the worker thread (after
a memcpy transfer) and the comm. thread (after a CMA or an RDMA transfer).
This caused an occasional hang because the condition (counter == numChildren)
was not satisfied. This patch makes 'counter' an atomic variable for SMP
modes to avoid the unsafe access and hence fixes the hang.

Change-Id: I8e77327da3682c86b3f952ba27957cf2c363a7d3