AMPI: Remove fsglobals and pipglobals from automated testing
Several autobuild targets have been failing in these tests because older
platforms and toolchain versions are not as forgiving about loading a
binary as a shared object without explicitly building it as one. A
partial solution is possible by overhauling our build scripts to make
this happen, but it requires intensive and potentially fragile changes
to the build process of user applications, and at present the methods
would break entirely if Fortran module variables are present, with no
known workaround.