bugfix (#1473) : fix ancient ordering bug in benchmark
Test has always incorrectly ordered the local call
to operationFinished in the timed allocation case
to be before the remote message send. Whereas it should
always be after, as it is in the untimed allocation case.
Bug revealed by Phil's introduction of size testing in 2017.
Falsely attributed to verbs, as this bug is invariant
with network layer.