Make group bcasts node aware in charm
To achieve this, we added a message type for broadcasts to groups, and only
call node level versions of converse broadcasts. When the message is received
it then uses a within-node broadcast on the charm side to forward to the rest
of the PEs on a node. This allows us to respect [nokeep] and prevent unneeded
message copies.