NO-MERGE feature (#1485): enable CharmDebug for SMP
This patch modifies netlrts locking behavior to
interleave CCS handling with normal launch.

Note: only applies to netlrts, but charmdebug has never supported
any other layers.  Unclear if this change is 100% correct, but no
correctness violations have been uncovered in testing so far.

This is only the tip of the iceberg of changes necessary for
complete SMP charmdebug support.