Charm4py: support new sections implementation
This adds support for a new sections implementation in Charm4py,
which I have tested internally, but would like other users to try
it. It doesn't interfere with existing functionality.

This removes Charm++ sections-related code from ckcallback when
building with CMK_CHARMPY. The rest of the changes only affect
CMK_CHARMPY code.