Change PICS configuration method to use PICS_configure
Previously, each PICS setting had its own function. However, this led to synchronization issues
due to the step starting before all the configuration was done. This change adds a centralized
PICS_configure function which takes in a struct PicsConfig which contains all the required
settings, and a callback which will be called once the configuration is done.