Add support to access ckLocalBranch from a node-local PE
This commit adds support for ckLocalBranchOther which can
be used to access a group from a node-local PE by passing a rank.
This support was required for incrementing the 'bcastNo' from the
comm. thread, without any involvement from the PE 0 worker thread.
This capability is useful for ZC Bcast when the broadcast root node
is non-zero.