bugfix 1999: record/replay may hang in SMP
envelopeEventID is the ckpv that monotonically increases
with each envelope (and therefore message).  However, in SMP
the envelope allocated and enqueued in the initDone process
is invisible to the record/replay process and can occur
on any PE in the node.  That creates a non-deterministic
race condition in the PE:envelopeEventID tuple space.

This workaround simply avoids incrementing the envelopeEventID
on the PE that triggers doneInit. This allows otherwise
deterministic SMP applications to record/replay with proper
fidelity.