ZC p2p CMA: Unpup/Pup rootNode in CkRdmaPrepareP2PMsg
The bug was introduced by the patch that supported Bcast from
a non-zero root. That patch required 'rootNode' to be pupped/unpupped
after 'numops' in a ZC metadata message. CkRdmaPrepareP2PMsg
opens up the ZC metadata message to modify it for supporting
de-registration in P2P CMA transfers and requires the additional
'rootNode' to be pupped/unpupped for correct processing of the
metadata message.