bugfix (#1485): fix charmdebug breakpoints in SMP mode
Several places in cpd startup constrained initialization to
CmiMyRank==0 based on defunct non-smp assumptions. The
restrictions for CmiRank 0 in cpd and ccs have been cut down
to the minimum.

Likewise the cmiArgDebugFlag has been refined to only
indicate that the debugger is attached. Code that formerly relied
on the (now false) assumption that it was only set on CmiRank 0
now tests for that condition explicitly. +cpd is now parsed
earlier to set cmiArgDebugFlag early enough for CmiTimerInit
to use it.

The barriers in CmiTimerInit caused a startup hang when the
debugger is attached.  Those are now disabled by cmiArgFlagDebug.
This does mean that timers may slip slightly when charmdebug
is attached, however, that is unlikely to create trouble as the
perturbations from charmdebug itself will be many orders of magnitude
greater than that slip.

atomic<int> breakPointSet has been added to the EntryInfo struct
(when charmdebug is enabled) to safely handle entry table
modifications for entry point breakpoints.

This also includes a bunch of commented out print statements,
because debugging a debugger leaves you with a dependency on
bare bones tools.