Add TreeLB and several lb strategies dependent on it
TreeLB is meant to replace CentralLB and HybridBaseLB. It does
centralized load balancing when configured as a 2-level tree.
It can work with multiple load balancing strategies, which can be
implemented in a separate source file and then added to
strategy_factory.h

This replaces the old common lb strategies like Greedy,
GreedyRefine and Refine with their new TreeLB-based
implementations.