Redesign Isomalloc to better meet AMPI's needs (#2524)
Originally, each PE was assigned a section of the global address space
and was responsible for managing it. The new paradigm divides the
available address space by logical units of work, such as AMPI virtual
ranks. This way, each rank can manage itself independently of all others
once memory ranges are assigned, and a rank's underlying memory will
always be contiguous in ownership allowing it to be migrated as a single
buffer.

The nature of Isomalloc requires it to interface with the system using
mmap, directly affecting virtual memory at the level of pages. This
means some form of pooling is required to avoid wasting memory when
servicing malloc calls smaller than the system page size, as well as to
minimize the number of expensive system calls made during runtime. The
new design includes the Isomempool to achieve these goals. Previously
each allocation made using Isomalloc would result in its own call to
mmap, which wasted memory space and computing time. The GNI machine
layer's mempool had also been available to Isomalloc, but it suffered
from correctness and efficiency issues, and unclear code. These two
modes of operation have been removed entirely in favor of the
Isomempool.

The Isomempool uses a red-black tree to store the locations and sizes of
free space within the region provided by the system through mmap. This
tree is used to determine the best placement for user allocations. Each
user allocation is prefixed with a header that points to its immediate
neighbors, making a doubly linked list of all regions. Deallocations
mark a region available but leave its header in place so that an
adjacent deallocation can merge the two empty regions.

This patch also cleans up tlsglobals and swapglobals due to the close
ties between migratable threads and global variable privatization. This
includes their API, how they are initialized, and when they are
(de)activated.