Make group broadcasts node aware in charm (#2440)
* Reformat _skipCldEnqueue

* Make SendToPeers respect nokeep

* Make group bcasts node aware in charm

To achieve this, we added a message type for broadcasts to groups, and only
call node level versions of converse broadcasts. When the message is received
it then uses a within-node broadcast on the charm side to forward to the rest
of the PEs on a node. This allows us to respect [nokeep] and prevent unneeded
message copies.

* Add group broadcast testing to WNB tests

* Use new CkAbort API in WNB test

* Update msg enum check test

* Update CHARMDEBUG_MINOR

* Changes to get ZC API working with within node bcast semantics

* Fix error while guarding SendToPeers in machine-broadcast.C

* Reset bcast root for msg forwarding on intermediate nodes for bcast post api

This commit also removes the change that was limiting the setting
of BocBcastMsg to just SMP mode, and this turn fixes the hang that
was seen on uth builds.

* Stop BIGSIM builds from using new bcast message type

* Add comments to describe BIGSIM workaround

Co-authored-by: Nitin Bhat <nitin.bhat.k@gmail.com>