Add a pup_buffer API with ZC functionality (#2458)
* Template methods in pup.h

* Add PUP::er::pup_buffer methods to allow flexible pupping/unpupping of buffers

The pup_buffer methods allow users to invoke a 'deallocate' method
to free the passed buffer, after packing as bytes and similarly invoke
an 'allocate' method to allocate a new buffer and assign it to the
passed buffer before unpacking as bytes. There is a default version,
where 'free' and 'malloc' will be called when the user doesn't pass
allocate and deallocate methods. The pup_buffer methods have the
following signature:

  void pup_buffer(T *&a, size_t nItems);

  void pup_buffer(T *&a, size_t nItems,
                        std::function<void *(size_t)> allocate,
                        std::function<void (void *)> deallocate);

This API was ideated by Evan Ramos from Charmworks, Inc.

* Add a pup_buffer API test

* Create CmiNcpyBuffer to pull out generic ZC code from CkNcpyBuffer

* ZC Pup: Add zerocopy functionality to the pup_buffer API

* Print typeid(*a).name() in error message

* Comments in pup.h about standard and custom pup_buffer methods

* Minor changes to pup_buffer.C test

* Changes to get ZC Pup to work on CMK_REG_REQUIRED layers

* Fixing compilation errors on generic layer

* Freeing allocated objects

* Add inline function in conv-rdma.h

* Waiting for Rgets before calling ResumeFromSync and ckJustMigrated

* Cleanup

* Getting smp mode to work

* Keeping Bigsim happy by renaming Ckpv variables as Cpv

* Delete buffer inside toDisk::pup_buffer only when isDeleting is true

This fixes the crash that was seen previously while running tests/ampi/chkpt
with Isomalloc modified to call pup_buffer.

* Addressing review comments (eliminating runtime branch cost)

* Make pup_buffer test more robust

* Support non-migration layers in pup_buffer test

Layers like Uth and Sim-Linux that have CMK_LBDB_ON set to 0, don't
support migration, causing `ckJustMigrated` to not be called. This is
fixed by appropriately modifying the test to accommodate such layers.