Fix verbs on new InfiniBand hardware/drivers
Verbs started rejecting setting retry_cnt to 20 because this is only
a 3-bit field, so the maximum legal value is 7.  This caused a
retry using "QLOGIC" values, which attempted to modify the mtu,
but the mtu flag was not set and mtu could not be modified at this
point of setup anyway, so the value was only changed for future
nodes attempting to connect and was in general inconsistent,
resulting in "Work completion error in sendCq" Charm++ aborts
along with mlx5 driver "got completion with error" messages.

Fix is to always set retry_cnt to 7 and don't try to change mtu.