cdash/ctest: Fix test directories on Windows (#2802)
On Windows systems, ./build does not create the usual top-level bin, tmp, etc. symlinks, which ctest used to find the directory of tests to run. This patch changes this to determine the name of the directory where the include/VERSION file is located, and use that directory to locate the tests/, examples/, and benchmarks/ directories.

Note that this might fail when there are multiple builds within the same charm directory, as there will be multiple include/VERSION files. Since this is meant just for automated test builds, we can ignore this limitation for now.