Fix rare bug seen in ZC EM Post API examples
The bug was causing a segfault because of the execution of the POST
entry method before the allocation of the receive buffers. This was
because of the buffer not being allocated in the object's
constructor (but another entry method). The fix involved moving the
allocation of the receive buffers to the object's constructor. This ensured
that the buffers are allocated before the POST entry method executes.