Add EndReceiver to CkSyncBarrier
This allows users to add a callback to run after the normal AtSync
calbacks have finished, which can be used for cleanup or to prepare
for resuming normal execution. Namely, this is used by CkLocMgr in
order to tell the LB framework that it has stopped inserting objects,
which was previously done by shoehorning it into being an AtSync
client unnaturally.