Fix hang in AtSync flooding mechanism
The flooding mechanism of AtSync, intended to cover cases where there
are empty PEs, can hang when the flood message is received by a PE
before that PE has exited its AtSync section, meaning that its
CkSyncBarrier is off. To fix this, track the epoch of the flood
message (now called "kick" to more closely match the semantics of the
procedure) so the barrier can check if the kick for the current epoch
has been received when the barrier is eventually enabled again.