AtSync: Change kick epoch semantics to match regular AtSync elements
All AtSync elements on a PE must have an epoch greater than the
current epoch of the CkSyncBarrier in order to trigger the
barrier. Previously, the kick epoch had to be greater than or equal to
the CkSyncBarrier epoch to trigger it for PEs without AtSync elements,
now it must strictly be greater to trigger the barrier, matching the
behavior of regular AtSync elements.