AtSync: Wait for all receivers to finish before resuming clients
Previously, as soon as any receiver called resumeClients, the
CkSyncBarrier would resume its clients. When there are multiple
receivers, this may be premature, as another receiver may be expecting
that the clients have not yet resumed. This change makes the barrier
wait for all receivers to call resumeClients before actually resuming
its clients.