Collide: Add support for 2-phase reduction
The first phase is a voxel reduction to indication completion.
This phase doesn't involve concatenation of the collision results.
When the callback of the first phase is called, the user can start
the second phase by calling getCollisionResults, which initiates
the second voxel reduction with the collision results executing
the callback of the second phase.