Use new {Begin,End}Receiver for CkLocMgr with AtSync
CkLocMgr was shoehorned to act as an AtSync client because it needs to
call DoneRegisteringObjects when load balancing finishes and there was
previously no other way to do this. This involved special casing it in
the CkSyncBarrier code because it wasn't a real client. This change
removes that hack. Additionally, this decouples AtSync from the
existence of an LB, allowing it to be easily used for non-LB purposes.