Do not skip commbench on Golub (#3281)
The previously seen slowdown in commbench was a result of incorrect
PE-binding caused because of not specifying the number of processes
in the slurm job script. Since it was not a bug within charm, this
commit reverts the skipping of benchmarks/converse/commbench since
it runs without any slowdown when the -n <processes> SBATCH directive
is passed in the slurm script.