Reimplement CmiNode{All}Barrier as spinbarrier (#3283)
The current implementation uses a condition variable for coordination
to avoid blocking for user level threads, etc. This works well at
scale, but can be very slow; this change replaces it with a
spinbarrier that progressively spins less aggressively the longer it
waits, benchmarking has shown that the new version has equal to or
better performance than the previous version.