ZC MultiplePostedBuffers Example: Fix incorrect callback creation
This patch fixes the crash seen in this example on UCX CI builds.
Previously, the compReductionCb was constructed in the constructor
of zerocopyObject when the mainProxy datamember was uninitialized.
With this patch, the compReductionCb construction call has been moved
to the testZeroCopy method, where the mainProxy is passed as a
parameter.