Trace communication to inline entry methods for LB (#3352)
Previously, messages sent to inline entry methods were not recorded in
the comm graph, causing comm oriented LBs to completely ignore that
communication. This change fixes that, both for when inline EPs are
actually local and called via the object pointer and when they are
remote and called via the normal messaging stack.