OFI: Change return type of registerDirectMemory to void
This commit fixes the ofi autobuild crashes seen on Stampede2.
Previously, registerDirectMemory was defined to return a fid_mr
struct pointer but did not return any argument. This
caused a mysterious crash with SIGSEGV on Stampede2 with gcc 9.1.0.
This commit changes the return type to void, which is the correct
return type for that function, and this in turn fixes the crash.