ZC Post API: Make size updation uniform on all elements (#3514)
Previously, the size param being passed into the regular entry method
was reflected to have the size posted by the user. However, this was
only done for the secondary elements i.e. elements which do not issue
the Get call. However, the primary element i.e. the element issuing the
Get call, always used the sender passed size parameter. With this change,
all elements only use the sender passed size, since it is odd to change
the value of a non-pointer type like size and this behavior deviates from
what happens in regular Charm++ entry methods.