ZC Post API: Set pointer to nullptr before passing it to Post EM
Previously, the pointer was set to the source buffer's pointer, which
is invalid when used directly on another node. For this reason, this
commit sets it to nullptr. Using the nullptr value, this commit also
adds error checking to ensure that the CkMatchBuffer call was made or
the reference to the pointer was set inside the Post Entry Method.