Use std::vector::swap and pass argument as pointers
This commit brings two major changes:
1. Previously, the vectors were swapped using 3 move operations
in favor of keeping the code changes to a minimal. This commit
switches to a more traditional std::swap specialization for
vectors a.k.a std::vector<T, A>::swap. This reduces the total
copy of pointers to 9 words as opposed to previous 18 words.
While this is no performance gain, this use ensures better
readability.
2. The switch from sending std::vector as entry method argument
to pointers to avoid unnecessary/unavoidable copies as pointed
out by Justin in his review.