Add option to enforce constraints during migration
When constraining things like memory footprint, the amount of memory
used may exceed what is assigned to an element during migration, since
it is possible for all old objects + all incoming new objects to be
resident at one time. This commits adds a "constrain_during_migration"
option to the configuration to allow users to enable this for specific
constraints.