feature: support CXI in OFI for Slingshot-11
This supports the CXI interface for Cassini (AKA Slingshot-11)
within the OFI machine layer.  No application level changes
are required, but a variety of command line options are
provided to configure the memory pool and the selection
of cxi interfaces.

Note: this does require the use of the memory pool in order to
efficiently support the FI_MR_ENDPOINT mode of memory registration.
As the time cost of registering individual messages would be
otherwise be entirely too high.

Charmrun has been configured to wrap srun and currently
assumes PMI2 with cray extensions for launching.

The build system has been set up to autodetect CXI and enable
support for it accordingly.  For compatibility purposes, it also
supports the use of cxi on the build line, but that should not be
necessary on most HPE systems with proper LMOD environments.