feature: support CXI in OFI for Slingshot-11 (#3791)
* feature: support CXI in OFI for Slingshot-11

This supports the CXI interface for Cassini (AKA Slingshot-11)
within the OFI machine layer.  No application level changes
are required, but a variety of command line options are
provided to configure the memory pool and the selection
of cxi interfaces.

Note: this does require the use of the memory pool in order to
efficiently support the FI_MR_ENDPOINT mode of memory registration.
As the time cost of registering individual messages would be
otherwise be entirely too high.

Charmrun has been configured to wrap srun and currently
assumes PMI2 with cray extensions for launching.

The build system has been set up to autodetect CXI and enable
support for it accordingly.  For compatibility purposes, it also
supports the use of cxi on the build line, but that should not be
necessary on most HPE systems with proper LMOD environments.

* !fixup: default opt_build_ofi to 0

* !fixup: autodetect and set CMK_CXI in build

* !fixup: reorder fields in OFIRmaHeader to reduce pad space

* cleanup: use DETERMINE_OFFSET macro for onesided

* update xpmem to use CMK_CXI instead of defunct CMK_OFI_CXI

this allows xpmem to be supported in the build process for nonsmp
targets.  This has not yet proven beneficial for performance.

* !cleanup: update comment to reflect xpmem not on by default

* !cleanup: update comment to reflect xpmem not on by default

* backwards compatibility changes to non-CXI OFI use cases

* !fixup: replace defunct CMK_OFI_CXI with CMK_CXI

* cleanup: revise default memory pool settings to be leaner

* !fixup: add comment regarding default off for CMK_CXI

* !cleanup: revise comment to match code
enable thread diagnostics in debug case

* !fixup: use rbuf consistently instead of rmsg

* doc: add description of cxi build option

---------

Co-authored-by: Eric Bohm <ebohm@login2.crusher.olcf.ornl.gov>
Co-authored-by: Eric J Bohm <ebohm@illinois.edu>